First pass at getting an AuthorsField
This starts down the path of getting an AuthorsField that pulls in all
of the behavior expected of Authors.  This is meant as a direct
comparison to the `AuthorshipMixin` class in @niran's fork and has that
behavior as expected.

Using an intermediate model means you can get all of the times a given
`User` has been attached to a model as one of its `Authors`, but it also
means that you have to do N+1 queries to get at all of those models.