Add extra logic around spacing of `extra`
This is needed to handle cases where the `extra` field has a comma, or
some other character that needs to be added without a space so it is
directly against the end of the generated string.

Note: This deviates from @niran's implementation is that it doesn't use
a &nbsp; as a separator in the case of HTML.  1) this code isn't aware
of what format it's going to generate; 2) there are plenty of cases
where you might want this to break on the space.