Move code toward its more permanent home
Going to try to break the arm_content code up into smaller chunks so it
could eventually be broken into multiple apps.  I like the idea of
having each of these as separate, but for the time being we'll leave
them like this as part of one big repo with a lot of sub-modules.