Remove a bunch of cruft from really old prototyping
None of this is needed any longer.  Originally, we didn't have any
concrete implementations of a content class.  For testing purposes we
can assume that `armstrong.apps.content` is around (for now), so we're
relying on it.

The potential problem with this commit is for people who attempt to test
`arm_content` via Django's test runner without having the main
`armstrong.apps.content` app installed.  Opening a ticket to deal with
that.